requirements.txt
aiogram==2.25
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
lxml==5.1.0
magic-filter==1.0.12
multidict==6.0.4
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
urllib3==2.1.0
yarl==1.9.4